% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asthma}
\alias{asthma}
\alias{measurements}
\alias{classes}
\title{Asthma RNA Abundance and Patient Classes}
\format{
\code{measurements} has a row for each sample and a column for each
gene. \code{classes} is a factor vector with values No and Yes, indicating if
a particular person has asthma or not.
}
\source{
A Nasal Brush-based Classifier of Asthma Identified by Machine
Learning Analysis of Nasal RNA Sequence Data, \emph{Scientific Reports},
2018.  Webpage: \url{http://www.nature.com/articles/s41598-018-27189-4}
}
\description{
Data set consists of a matrix of abundances of 2000 most variable gene
expression measurements for 190 samples and a factor vector of classes for
those samples.
}
\keyword{datasets}
