\name{asthma}
\docType{data}
\alias{measurements}
\alias{classes}
\title{Asthma RNA Abundance and Patient Classes}
\description{
  Data set consists of a matrix of abundances of 2000 most variable gene expression measurements for 190 samples and a factor vector of classes for those samples.
}
\usage{data(asthma)}
\format{\code{measurements} has a row for each gene and a column for each sample. \code{classes} is a factor vector.}
\source{A Nasal Brush-based Classifier of Asthma Identified by Machine Learning Analysis of Nasal RNA Sequence Data,
       \emph{Scientific Reports}, 2018.
       Webpage: \url{http://www.nature.com/articles/s41598-018-27189-4}}
\keyword{datasets}