% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingBartlett.R
\name{bartlettRanking}
\alias{bartlettRanking}
\alias{bartlettRanking,matrix-method}
\alias{bartlettRanking,DataFrame-method}
\alias{bartlettRanking,MultiAssayExperiment-method}
\title{Ranking of Differential Variability with Bartlett Statistic}
\usage{
\S4method{bartlettRanking}{matrix}(measurementsTrain, classesTrain, ...)

\S4method{bartlettRanking}{DataFrame}(measurementsTrain, classesTrain, verbose = 3)

\S4method{bartlettRanking}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  classesTrain,
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information table will be
used.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Ranks all features from largest Bartlett statistic to smallest.
}
\details{
The calculation of the test statistic is performed by the
\code{\link{bartlett.test}} function from the \code{\link{stats}} package.

Data tables which consist entirely of non-numeric data cannot be ranked.
}
\examples{

  # Samples in one class with differential variability to other class.
  # First 20 genes are DV.
  genesRNAmatrix <- sapply(1:20, function(sample) c(rnorm(25, 9, 1), rnorm(25, 9, 5)))
  genesRNAmatrix <- cbind(genesRNAmatrix, sapply(1:80, function(sample) rnorm(50, 9, 1)))
  rownames(genesRNAmatrix) <- paste("Sample", 1:50)
  colnames(genesRNAmatrix) <- paste("Gene", 1:100)
  genesSNPmatrix <- matrix(sample(c("None", "Missense"), 250, replace = TRUE), nrow = 50)
  rownames(genesSNPmatrix) <- paste("Sample", 1:50)
  colnames(genesSNPmatrix) <- paste("Gene", 1:5)
  classes <- factor(rep(c("Poor", "Good"), each = 25))
  names(classes) <- paste("Sample", 1:50)
  
  
  genesDataset <- MultiAssayExperiment(list(RNA = t(genesRNAmatrix), SNP = t(genesSNPmatrix)),
                                       colData = DataFrame(class = classes))
                                         
  bartlettRanking(genesDataset, targets = "RNA", classesTrain = "class")

}
\author{
Dario Strbenac
}
