\name{BiocParallelParamOrMissing}
\docType{class}
\alias{BiocParallelParamOrMissing}
\alias{BiocParallelParamOrMissing-class}

\title{Union of BiocParallelParam and missing}
\description{
  Allows a method parameter to be either a BiocParallelParam class object or missing.
}

\author{Dario Strbenac}
