% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraPlots.R
\name{Boxplot}
\alias{Boxplot}
\title{Boxplot for looking at performance or ClassifyR object}
\usage{
Boxplot(
  result,
  metric = "Balanced Accuracy",
  x = "`Classifier Name`",
  fill = "dataset"
)
}
\arguments{
\item{result}{A list of ClassifyR objects}

\item{metric}{A performance metric from \code{calcPerformance}}

\item{x}{Which Characteristic to use for the x-axis}

\item{fill}{Which Characterstic to use for the fill}
}
\value{
ggplot object
}
\description{
Boxplot for looking at performance or ClassifyR object
}
\examples{

data(asthma)

# Compare randomForest and svm classifiers.
result <- crossValidate(measurements, classes, nRepeats = 4)
Boxplot(result)

}
