% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{bundle_model}
\alias{bundle_model}
\title{Bundle model components into a classifyr object}
\usage{
bundle_model(preprocessor, model, target, features)
}
\arguments{
\item{preprocessor}{First stage of our model.}

\item{model}{Second stage of our model.}

\item{target}{The target column used in training.}

\item{features}{The feature column names used in training.}
}
\value{
A classifyr_model object.
}
\description{
Bundle model components into a classifyr object
}
\keyword{internal}
