\name{characterOrDataFrame}
\docType{class}
\alias{characterOrDataFrame}
\alias{characterOrDataFrame-class}

\title{Union of a Character and a DataFrame}
\description{
  Allows a slot to be either a character or a DataFrame.
}

\author{Dario Strbenac}
\examples{
  # Used for the originalFeatures slot of a ClassifyResult object.
}
