% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyr-package.R
\docType{package}
\name{classifyr-package}
\alias{classifyr}
\alias{classifyr-package}
\title{classifyr: Example Classification Model}
\description{
This package skeleton supplements the posit::conf 2025 talk “Building Governable ML Models in R” and demonstrates several useful patterns for R-based modeling projects.
}
\author{
\strong{Maintainer}: Tom Shafer \email{tom.shafer@elderresearch.com}

Other contributors:
\itemize{
  \item Elder Research [copyright holder]
}

}
\keyword{internal}
