% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ClassifyResult}
\alias{ClassifyResult}
\alias{ClassifyResult-class}
\alias{ClassifyResult,DataFrame,character,characterOrDataFrame-method}
\alias{show,ClassifyResult-method}
\alias{sampleNames}
\alias{sampleNames,ClassifyResult-method}
\alias{featureNames}
\alias{featureNames,ClassifyResult-method}
\alias{predictions}
\alias{predictions,ClassifyResult-method}
\alias{actualOutcomes}
\alias{actualOutcomes,ClassifyResult-method}
\alias{features}
\alias{features,ClassifyResult-method}
\alias{models}
\alias{models,ClassifyResult-method}
\alias{performance}
\alias{performance,ClassifyResult-method}
\alias{tunedParameters}
\alias{tunedParameters,ClassifyResult-method}
\alias{totalPredictions}
\alias{totalPredictions,ClassifyResult-method}
\title{Container for Storing Classification Results}
\usage{
ClassifyResult(characteristics, originalNames, originalFeatures, ...)
}
\description{
Contains a list of models, table of actual sample classes and predicted
classes, the identifiers of features selected for each fold of each
permutation or each hold-out classification, and performance metrics such as
error rates.  This class is not intended to be created by the user. It is
created by \code{\link{runTest}} or \code{\link{runTests}}.
}
\section{Constructor}{
 \describe{ \item{}{
\preformatted{ClassifyResult(characteristics, originalNames,
originalFeatures, rankedFeatures, chosenFeatures, predictions,
actualOutcomes, models, validation, tune = NULL)} } } \describe{
\item{list("characteristics")}{A \code{\link{DataFrame}} describing the
characteristics of classification done. First column must be named
\code{"charateristic"} and second column must be named \code{"value"}. If
using wrapper functions for feature selection and classifiers in this
package, the function names will automatically be generated and therefore it
is not necessary to specify them.} \item{list("originalNames")}{All sample
names.} \item{list("originalFeatures")}{All feature names. Character vector
or \code{\link{DataFrame}} with one row for each feature if the data set is
a \code{\link{MultiAssayExperiment}}.} \item{list("rankedFeatures")}{All
features, from most to least important. Character vector or
\code{\link{DataFrame}} if data set is a \code{MultiAssayExperiment}.}
\item{list("chosenFeatures")}{Features selected at each fold. Character
vector or DataFrame if data set is a \code{MultiAssayExperiment}.}
\item{list("predictions")}{A \code{\link{list}} of \code{\link{data.frame}}
containing information about samples, their actual class and predicted class
and/or class score and information about the cross-validation fold in which
the prediction was made.} \item{list("actualOutcomes")}{Factor of class of
each sample.} \item{list("models")}{All of the models fitted to the training
data.} \item{list("validation")}{List with first element being the name of
the validation scheme, and other elements providing details about the
scheme.} \item{list("tune")}{A description of the tuning parameters, and the
value chosen of each parameter.} }
}

\examples{

  #if(require(sparsediscrim))
  #{
    data(asthma)
    
    LOOCVparams <- CrossValParams("Leave-k-Out", leave = 1)
    modellingParams <- ModellingParams()
    classified <-
    runTests(measurements, classes, LOOCVparams, modellingParams,
             DataFrame(characteristic = c("dataset", "classification"),
                      value = c("Asthma", "Different Means"))
             )
    class(classified)
  #}
  
}
\author{
Dario Strbenac
}
