% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceCoxnet.R
\name{coxnetTrainInterface}
\alias{coxnetTrainInterface}
\alias{coxnetInterface}
\alias{coxnetPredictInterface}
\alias{coxnetTrainInterface,matrix-method}
\alias{coxnetTrainInterface,DataFrame-method}
\alias{coxnetTrainInterface,MultiAssayExperiment-method}
\alias{coxnetPredictInterface,multnet,matrix-method}
\alias{coxnetPredictInterface,multnet,DataFrame-method}
\alias{coxnetPredictInterface,multnet,MultiAssayExperiment-method}
\title{An Interface for glmnet Package's coxnet Function}
\usage{
coxnetTrainInterface(measurementsTrain, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method (e.g. \code{verbose}) or, for the training function,
options that are used by the \code{glmnet} function. For the testing
function, this variable simply contains any parameters passed from the
classification framework to it which aren't used by glmnet's \code{predict}
fuction.}

\item{survivalTrain}{A tabular data type of survival information of the
same number of rows as the number of samples in \code{measurementsTrain} and 2 to 3 columns if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}}, or a character vector of length 2 to 3 containing the
column names in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a vector of column names, those columns will be
removed before training.}

\item{lambda}{The lambda value passed directly to
\code{\link[glmnet]{glmnet}} if the training function is used or passed as
\code{s} to \code{\link[glmnet]{predict.glmnet}} if the prediction function
is used.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurements} and the same number of features
as it.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"clinical"} is also a valid value
and specifies that integer variables from the clinical data table will be
used.}

\item{model}{A trained coxnet, as created by the \code{glmnet}
function.}

\item{returnType}{Default: \code{"both"}. Either \code{"class"},
\code{"score"} or \code{"both"}.  Sets the return value from the prediction
to either a vector of class labels, matrix of scores for each class, or both
labels and scores in a \code{data.frame}.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
For \code{coxnetTrainInterface}, an object of type
\code{glmnet}. For \code{coxnetPredictInterface}, a matrix of containing the link and risk functions.
\code{returnType}.
}
\description{
An elastic net GLM classifier uses a penalty which is a combination of a
lasso penalty and a ridge penalty, scaled by a lambda value, to fit a sparse
linear model to the data.
}
\details{
The value of the \code{family} parameter is fixed to \code{"cox"} so
that classification with survival is possible.
During classifier training, if more than one lambda value
is considered by specifying a vector of them as input or leaving the default
value of NULL, then the chosen value is determined based on classifier
resubstitution error rate.
}
\examples{

  
}
