% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingCoxPH.R
\name{coxphRanking}
\alias{coxphRanking}
\alias{coxphRanking,matrix-method}
\alias{coxphRanking,DataFrame-method}
\alias{coxphRanking,MultiAssayExperiment-method}
\title{Ranking of Differential Variability with coxph Statistic}
\usage{
coxphRanking(measurementsTrain, ...)

\S4method{coxphRanking}{matrix}(measurementsTrain, survivalTrain, ...)

\S4method{coxphRanking}{DataFrame}(measurementsTrain, survivalTrain, verbose = 3)

\S4method{coxphRanking}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  survivalTrain,
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method.}

\item{survivalTrain}{A tabular data type of survival information of the
same number of rows as the number of samples in \code{measurementsTrain} and 2 to 3 columns if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}}, or a character vector of length 2 to 3 containing the
column names in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a vector of column names, those columns will be
removed before training.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"clinical"} is also a valid value
and specifies that numeric variables from the clinical data table will be
used.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Ranks all features from largest coxph statistic to smallest.
}
\details{
The calculation of the test statistic is performed by the
\code{\link{coxph.test}} function from the \code{\link{stats}} package.

Data tables which consist entirely of non-numeric data cannot be ranked.
}
