% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DataFrameOrNULL-class}
\alias{DataFrameOrNULL-class}
\alias{DataFrameOrNULL}
\title{Union of NULL and DataFrame Class}
\description{
Allows cross-validation to accept data as either a \code{DataFrame} (for a
single data set) or \code{DataFrameList} (for a list of tables of related
measurements, such as different projects measuring the same outcome and the
same kind of measurements). No constructor.
}
\author{
Dario Strbenac
}
