% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain a model's predictions}
\usage{
explain(model, preds, data)
}
\arguments{
\item{model}{The model object to explain.}

\item{preds}{A vector of model predictions.}

\item{data}{The source data leading to the predictions.}
}
\description{
Explain a model's predictions
}
