\name{FeatureSetCollectionOrNULL}
\docType{class}
\alias{FeatureSetCollectionOrNULL}
\alias{FeatureSetCollectionOrNULL-class}

\title{Union of a FeatureSetCollection and NULL}
\description{
  Allows a slot to be either a FeatureSetCollectionOrNULL object or empty.
}

\author{Dario Strbenac}
\examples{
  TrainParams(DLDAtrainInterface, transform = NULL) # Use the input data as-is.
}
