% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FeatureSetCollectionOrNULL}
\alias{FeatureSetCollectionOrNULL}
\alias{FeatureSetCollectionOrNULL-class}
\title{Union of a FeatureSetCollection and NULL}
\description{
Allows a slot to be either a FeatureSetCollectionOrNULL object or empty.
}
\examples{

  TrainParams(DLDAtrainInterface, transform = NULL) # Use the input data as-is.

}
\author{
Dario Strbenac
}
