% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{functionOrList-class}
\alias{functionOrList-class}
\alias{functionOrList}
\title{Union of Functions and List of Functions}
\description{
Allows a slot to be either a function or a list of functions.
}
\examples{

  SelectParams(limmaRanking)
  SelectParams(list(limmaRanking, differentMeansRanking)) # Ensemble selection.

}
\author{
Dario Strbenac
}
