% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{functionOrNULL}
\alias{functionOrNULL}
\alias{functionOrNULL-class}
\title{Union of A Function and NULL}
\description{
Allows a slot to be either a function or empty.
}
\examples{

  PredictParams(NULL)
  PredictParams(DLDApredictInterface)

}
\author{
Dario Strbenac
}
