% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidate.R
\name{generateCrossValParams}
\alias{generateCrossValParams}
\title{A function to generate a CrossValParams object}
\usage{
generateCrossValParams(nRepeats, nFolds, nCores, selectionOptimisation)
}
\arguments{
\item{nRepeats}{A numeric specifying the the number of repeats or permutations to use for cross-validation.}

\item{nFolds}{A numeric specifying the number of folds to use for cross-validation.}

\item{nCores}{A numeric specifying the number of cores used if the user wants to use parallelisation.}

\item{selectionOptimisation}{A character of "Resubstitution", "Nested CV" or "none" specifying the approach used to optimise \code{nFeatures}.}
}
\value{
CrossValParams object
}
\description{
A function to generate a CrossValParams object
}
\examples{
CVparams <- generateCrossValParams(nRepeats = 20, nFolds = 5, nCores = 8, selectionOptimisation = "none")
}
