% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceGLM.R
\name{Ordinary GLM Interface}
\alias{Ordinary GLM Interface}
\alias{GLMinterface}
\alias{GLMtrainInterface}
\alias{GLMpredictInterface}
\alias{GLMtrainInterface,matrix-method}
\alias{GLMtrainInterface,DataFrame-method}
\alias{GLMtrainInterface,MultiAssayExperiment-method}
\alias{GLMpredictInterface,glm,matrix-method}
\alias{GLMpredictInterface,glm,DataFrame-method}
\alias{GLMpredictInterface,glm,MultiAssayExperiment-method}
\title{An Interface for Ordinary Generalised Linear Models (GLMs) with Binary Outcome}
\usage{
\S4method{GLMtrainInterface}{matrix}(measurementsTrain, classesTrain, ...)

\S4method{GLMtrainInterface}{DataFrame}(measurementsTrain, classesTrain, ..., verbose = 3)

\S4method{GLMtrainInterface}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  classesTrain,
  ...
)

\S4method{GLMpredictInterface}{glm,matrix}(model, measurementsTest, ...)

\S4method{GLMpredictInterface}{glm,DataFrame}(
  model,
  measurementsTest,
  returnType = c("both", "class", "score"),
  verbose = 3
)

\S4method{GLMpredictInterface}{glm,MultiAssayExperiment}(
  model,
  measurementsTest,
  targets = names(measurementsTest),
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method (e.g. \code{verbose}) or, for the training function,
options that are used by the \code{\link{glm}} function. For the testing
function, this variable simply contains any parameters passed from the
classification framework to it which aren't used by the \code{\link{predict.glm}}
function.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that integer variables from the sample information data table will be
used.}

\item{model}{A trained GLM, as created by the \code{glm} function.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurementsTrain} and the same number of features
as it.}

\item{returnType}{Default: \code{"both"}. Either \code{"class"},
\code{"score"} or \code{"both"}.  Sets the return value from the prediction
to either a vector of class labels, matrix of scores for each class, or both
labels and scores in a \code{data.frame}.}
}
\value{
For \code{GLMtrainInterface}, an object of type
\code{list}. For \code{GLMpredictInterface}, either a factor
vector of predicted classes, a matrix of scores for each class, or a table
of both the class labels and class scores, depending on the setting of
\code{returnType}.
}
\description{
No feature selection is done, so some should be done beforehand or otherwise the
elastic net GLM classifier should be used instead of this classifier.
}
\details{
The value of the \code{family} parameter is fixed to \code{binomal}.
}
\examples{

    # Genes 76 to 100 have differential expression.
    genesMatrix <- sapply(1:100, function(sample) rnorm(25, 9, 0.3))
    genesMatrix <- rbind(genesMatrix, t(sapply(1:25, function(sample)
                                      c(rnorm(75, 9, 0.3), rnorm(25, 14, 0.3)))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    rownames(genesMatrix) <- paste("Sample", 1:nrow(genesMatrix))
    colnames(genesMatrix) <- paste("Gene", 1:ncol(genesMatrix))
    
    CVparams <- CrossValParams("k-Fold")
    trainParams <- TrainParams(GLMtrainInterface)
    predictParams <- PredictParams(GLMpredictInterface)
    modParams <- ModellingParams(selectParams = NULL, trainParams = trainParams,
                                   predictParams = predictParams)
    classified <- runTests(genesMatrix[, 76:80], classes, CVparams, modParams)
                           
    classified <- calcCVperformance(classified, "Balanced Error")
    head(tunedParameters(classified))
    performance(classified)
}
\author{
Dario Strbenac
}
