% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HuRI}
\alias{HuRI}
\alias{interactors}
\title{Human Reference Interactome}
\format{
\code{interactors} is a \code{\link{Pairs}} object containing each
pair of interacting proteins.
}
\source{
A Reference Map of the Human Binary Protein Interactome,
\emph{Nature}, 2020.  Webpage:
\url{http://www.interactome-atlas.org/download}
}
\description{
A collection of 45783 pairs of protein gene symbols, as determined by the
The Human Reference Protein Interactome Mapping Project. Self-interactions
have been removed.
}
\keyword{datasets}
