\name{HuRI}
\docType{data}
\alias{interactors}
\title{Human Reference Interactome}
\description{
  A collection of 45783 pairs of protein gene symbols, as determined by the The Human Reference Protein Interactome Mapping Project. Self-interactions have been removed.
}
\usage{data(HuRI)}
\format{\code{interactors} is a \code{\link{Pairs}} object containing each pair of interacting proteins.}
\source{A Reference Map of the Human Binary Protein Interactome, \emph{Nature}, 2020.
       Webpage: \url{http://www.interactome-atlas.org/download}}
\keyword{datasets}