% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingKolmogorovSmirnov.R
\name{KolmogorovSmirnovRanking}
\alias{KolmogorovSmirnovRanking}
\alias{KolmogorovSmirnovRanking,matrix-method}
\alias{KolmogorovSmirnovRanking,DataFrame-method}
\alias{KolmogorovSmirnovRanking,MultiAssayExperiment-method}
\title{Ranking of Differential Distributions with Kolmogorov-Smirnov Distance}
\usage{
\S4method{KolmogorovSmirnovRanking}{matrix}(measurementsTrain, classesTrain, ...)

\S4method{KolmogorovSmirnovRanking}{DataFrame}(measurementsTrain, classesTrain, ..., verbose = 3)

\S4method{KolmogorovSmirnovRanking}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  classesTrain,
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or options which are accepted by the function
\code{\link{ks.test}}.}

\item{classesTrain}{Either a vector of class labels of class \code{\link{factor}}
of the same length as the number of samples in \code{measurementsTrain} or if the
measurements are of class \code{DataFrame} a character vector of length 1
containing the column name in \code{measurement} is also permitted.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}
}
\value{
A vector of feature indices, from the most promising features in the
first position to the least promising feature in the last position.
}
\description{
Ranks features from largest Kolmogorov-Smirnov distance to smallest.
}
\examples{

  # First 20 features have bimodal distribution for Poor class.
  # Other 80 features have normal distribution for both classes.
  set.seed(1984)
  genesMatrix <- sapply(1:20, function(feature)
                              {
                                randomMeans <- sample(c(8, 12), 25, replace = TRUE)
                                c(rnorm(25, randomMeans, 1), rnorm(25, 10, 1))
                              }
                       )
  genesMatrix <- cbind(genesMatrix, sapply(1:80, function(feature) rnorm(50, 10, 1)))
  classes <- factor(rep(c("Poor", "Good"), each = 25))

  ranked <- KolmogorovSmirnovRanking(genesMatrix, classes)
  head(ranked)                           

}
\author{
Dario Strbenac
}
