% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingKullbackLeibler.R
\name{KullbackLeiblerRanking}
\alias{KullbackLeiblerRanking}
\alias{KullbackLeiblerRanking,matrix-method}
\alias{KullbackLeiblerRanking,DataFrame-method}
\alias{KullbackLeiblerRanking,MultiAssayExperiment-method}
\title{Ranking of Differential Distributions with Kullback-Leibler Distance}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or options which are accepted by the function
\code{\link{getLocationsAndScales}}.}

\item{classesTrain}{Either a vector of class labels of class \code{\link{factor}}
of the same length as the number of samples in \code{measurementsTrain} or if the
measurements are of class \code{DataFrame} a character vector of length 1
containing the column name in \code{measurement} is also permitted.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Ranks features from largest Kullback-Leibler distance between classes to
smallest.
}
\details{
The distance is defined as \eqn{\frac{1}{2} \times \big( \frac{(location_1 -
location_2)^2}{scale_1^2} + \frac{(location_1 - location_2)^2}{scale_2^2} +
\frac{scale_2^2}{scale_1^2} + \frac{scale_1^2}{scale_2^2}\big)}{0.5 *
((location1 - location2)^2 / scale1^2 + (location1 - location2)^2 / scale2^2
+ scale1^2 / scale2^2 + scale2^2 / scale1^2)}

The subscripts denote the group which the parameter is calculated for.
}
\examples{

  # First 20 features have bimodal distribution for Poor class.
  # Other 80 features have normal distribution for both classes.
  genesMatrix <- sapply(1:20, function(feature)
                              {
                                randomMeans <- sample(c(8, 12), 25, replace = TRUE)
                                c(rnorm(25, randomMeans, 1), rnorm(25, 10, 1))
                              }
                       )
  genesMatrix <- cbind(genesMatrix, sapply(1:80, function(feature) rnorm(50, 10, 1)))
  classes <- factor(rep(c("Poor", "Good"), each = 25))
  
  ranked <- KullbackLeiblerRanking(genesMatrix, classes)
  head(ranked)
  
}
\author{
Dario Strbenac
}
