% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingLikelihoodRatio.R
\name{likelihoodRatioRanking}
\alias{likelihoodRatioRanking}
\alias{likelihoodRatioRanking,matrix-method}
\alias{likelihoodRatioRanking,DataFrame-method}
\alias{likelihoodRatioRanking,MultiAssayExperiment-method}
\title{Ranking of Differential Distributions with Likelihood Ratio Statistic}
\usage{
\S4method{likelihoodRatioRanking}{matrix}(measurementsTrain, classesTrain, ...)

\S4method{likelihoodRatioRanking}{DataFrame}(
  measurementsTrain,
  classesTrain,
  alternative = c(location = "different", scale = "different"),
  ...,
  verbose = 3
)

\S4method{likelihoodRatioRanking}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  classesTrain,
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or options which are accepted by the function
\code{\link{getLocationsAndScales}}.}

\item{classesTrain}{Either a vector of class labels of class \code{\link{factor}}
of the same length as the number of samples in \code{measurementsTrain} or if the
measurements are of class \code{DataFrame} a character vector of length 1
containing the column name in \code{measurement} is also permitted. Not used
if \code{measurementsTrain} is a \code{MultiAssayExperiment} object.}

\item{alternative}{Default: \code{c("different", "different")}. A vector of
length 2.  The first element specifies the location of the alternate
hypothesis.  The second element specifies the scale of the alternate
hypothesis.  Valid values in each element are "same" or "different".}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Ranks features from largest difference of log likelihoods (null hypothesis -
alternate hypothesis) to smallest.
}
\details{
Likelihood ratio test of null hypothesis that the location and scale are the
same for both groups, and an alternate hypothesis that is specified by
parameters. The location and scale of features is calculated by
\code{\link{getLocationsAndScales}}. The distribution fitted to the data is
the normal distribution.
}
\examples{

  # First 20 features have bimodal distribution for Poor class.
  # Other 80 features have normal distribution for both classes.

  genesMatrix <- sapply(1:20, function(feature)
                              {
                                randomMeans <- sample(c(8, 12), 25, replace = TRUE)
                                c(rnorm(25, randomMeans, 1), rnorm(25, 10, 1))
                              }
                       )
  genesMatrix <- cbind(genesMatrix, sapply(1:80, function(feature) rnorm(50, 10, 1)))
  classes <- factor(rep(c("Poor", "Good"), each = 25))

  ranked <- likelihoodRatioRanking(genesMatrix, classes)
  head(ranked)

}
\author{
Dario Strbenac
}
