% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingLimma.R
\name{limmaRanking}
\alias{limmaRanking}
\alias{limmaRanking,matrix-method}
\alias{limmaRanking,DataFrame-method}
\alias{limmaRanking,MultiAssayExperiment-method}
\title{Ranking of Differentially Abundant Features}
\usage{
limmaRanking(measurementsTrain, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or optional settings that are passed to
\code{\link[limma]{lmFit}}.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurements}.}

\item{targets}{Names of data tables to be combined into a single table and
used in the analysis.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Uses a moderated F-test with empirical Bayes shrinkage to rank
differentially expressed features based on differences of means. This means
it works when there are three or more classes.
}
\details{
This ranking method looks for changes in means and uses a moderated F-test
to do so.
}
\examples{

  #if(require(sparsediscrim))
  #{
    # Genes 76 to 100 have differential expression.
    genesMatrix <- sapply(1:25, function(sample) c(rnorm(100, 9, 2)))
    genesMatrix <- cbind(genesMatrix, sapply(1:25, function(sample)
                                 c(rnorm(75, 9, 2), rnorm(25, 14, 2))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    colnames(genesMatrix) <- paste("Sample", 1:ncol(genesMatrix))
    rownames(genesMatrix) <- paste("Gene", 1:nrow(genesMatrix))
    
    ranked <- limmaRanking(genesMatrix, classes)
    head(ranked)                           
  #}

}
\references{
Limma: linear models for microarray data, Gordon Smyth, 2005,
In: Bioinformatics and Computational Biology Solutions using R and
Bioconductor, Springer, New York, pages 397-420.
}
\author{
Dario Strbenac
}
