% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{listOrNULL}
\alias{listOrNULL}
\alias{listOrNULL-class}
\title{Union of a List and NULL}
\description{
Allows a slot to be either a list or a NULL.
}
\examples{

  setClass("EasyClassifier", representation(model = "listOrNULL"))
  classifier <- new("EasyClassifier", model = NULL) # Optimistic classifier.

}
\author{
Dario Strbenac
}
