\name{melanomaResponse}
\docType{data}
\alias{measurements}
\alias{classes}
\title{Melanoma Patient Protein Measurements in Blood Cells and Patient Classes}
\description{
  Data set consists of a matrix of relative abundances of 35 proteins for 20 samples and a factor vector of classes for those samples.
}
\usage{data(melanomaResponse)}
\format{\code{measurements} has a row for each protein and a column for each sample. \code{classes} is a named factor vector.}
\source{High-dimensional Single-cell Analysis Predicts Response to Anti-PD-1 Immunotherapy,
       \emph{Nature Medicine}, 2018.
       Webpage: \url{https://www.nature.com/articles/nm.4466}}
\keyword{datasets}