% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ModellingParamsOrNULL-class}
\alias{ModellingParamsOrNULL-class}
\alias{ModellingParamsOrNULL}
\title{Union of A ModellingParams Object and NULL}
\description{
Allows a slot to be either a ModellingParams class object or empty. No
constructor.
}
