% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceNSC.R
\name{NSCtrainInterface}
\alias{NSCtrainInterface}
\alias{NSCtrainInterface,matrix-method}
\alias{NSCtrainInterface,DataFrame-method}
\alias{NSCtrainInterface,MultiAssayExperiment-method}
\title{Interface for \code{pamr.train} Function from \code{pamr} CRAN Package}
\usage{
NSCtrainInterface(measurementsTrain, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or extra arguments passed to \code{\link[pamr]{pamr.train}}.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurementsTrain} and the same number of features
as it.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
A list with elements as described in \code{\link[pamr]{pamr.train}}.
}
\description{
Restructures variables from ClassifyR framework to be compatible with
\code{\link[pamr]{pamr.train}} definition.
}
\details{
This function is an interface between the ClassifyR framework and
\code{\link[pamr]{pamr.train}}.
}
\examples{

  if(require(pamr))
  {
    # Samples in one class with differential expression to other class for last 25 features.
    genesMatrix <- sapply(1:100, function(sampleColumn) c(rnorm(25, 9, 1)))
    genesMatrix <- rbind(genesMatrix, cbind(sapply(1:50, function(sampleColumn) rnorm(25, 9, 1)),
                                            sapply(1:50, function(sampleColumn) rnorm(25, 14, 1))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    
    NSCtrainInterface(genesMatrix, classes)
  }

}
\seealso{
\code{\link[pamr]{pamr.train}} for the function that was interfaced
to.
}
\author{
Dario Strbenac
}
