% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{numericOrNULL-class}
\alias{numericOrNULL-class}
\alias{numericOrNULL}
\title{Union of A Numeric Value and NULL}
\description{
Allows a slot to be either a numeric value or empty. No constructor.
}
\author{
Dario Strbenac
}
