% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.classifyr_model}
\alias{predict.classifyr_model}
\title{Generate predictions for a classifyr model}
\usage{
\method{predict}{classifyr_model}(object, newdata, ...)
}
\arguments{
\item{object}{A pretrained classifyr model.}

\item{newdata}{A data frame for making predictions.}

\item{...}{Unused arguments.}
}
\description{
Generate predictions for a classifyr model
}
\keyword{internal}
