% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.classifyr_model}
\alias{print.classifyr_model}
\title{Print a string representation for our model object}
\usage{
\method{print}{classifyr_model}(x, ...)
}
\description{
Print a string representation for our model object
}
\keyword{internal}
