% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{randomForest}
\alias{randomForest}
\alias{randomForest-class}
\alias{coxph}
\alias{coxph-class}
\title{Trained randomForest Object}
\description{
Enables S4 method dispatching on it.

Enables S4 method dispatching on it.
}
\author{
Dario Strbenac
}
