% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{train}
\alias{Surv}
\alias{DataFrame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{train}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{DataFrame}}}

  \item{survival}{\code{\link[survival]{Surv}}}
}}

