% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceRandomForestSurvival.R
\name{rfsrcTrainInterface}
\alias{rfsrcTrainInterface}
\alias{rfsrcInterface}
\alias{rfsrcPredictInterface}
\alias{rfsrcInterface,matrix-method}
\alias{rfsrcInterface,DataFrame-method}
\alias{rfsrcInterface,MultiAssayExperiment-method}
\alias{rfsrcPredictInterface,rfsrc,matrix-method}
\alias{rfsrcPredictInterface,rfsrc,DataFrame-method}
\alias{rfsrcPredictInterface,rfsrc,MultiAssayExperiment-method}
\alias{rfsrcTrainInterface,matrix-method}
\alias{rfsrcTrainInterface,DataFrame-method}
\alias{rfsrcTrainInterface,MultiAssayExperiment-method}
\title{An Interface for randomForestSRC Package's rfsrc random forest survival Function}
\usage{
\S4method{rfsrcTrainInterface}{matrix}(measurementsTrain, survivalTrain, ...)

\S4method{rfsrcTrainInterface}{DataFrame}(
  measurementsTrain,
  survivalTrain,
  mTryProportion = 0.5,
  ...,
  verbose = 3
)

\S4method{rfsrcTrainInterface}{MultiAssayExperiment}(
  measurementsTrain,
  targets = names(measurementsTrain),
  survivalTrain,
  ...
)

\S4method{rfsrcPredictInterface}{rfsrc,matrix}(model, measurementsTest, ...)

\S4method{rfsrcPredictInterface}{rfsrc,DataFrame}(model, measurementsTest, ..., verbose = 3)

\S4method{rfsrcPredictInterface}{rfsrc,MultiAssayExperiment}(
  model,
  measurementsTest,
  targets = names(measurementsTest),
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method (e.g. \code{verbose}) or options which are accepted
by the \code{\link[randomForestSRC]{rfsrc}} or \code{\link[randomForestSRC]{predict.rfsrc}} functions.}

\item{survivalTrain}{A tabular data type of survival information of the
same number of rows as the number of samples in \code{measurementsTrain} and 2 to 3 columns if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}}, or a character vector of length 2 to 3 containing the
column names in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a vector of column names, those columns will be
removed before training.}

\item{mTryProportion}{Default: 0.5. The proportion of features to try at each split.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that integer variables from the clinical data table will be
used.}

\item{model}{A trained rfsrc classifier, as created by
\code{rfsrcTrainInterface}, which has the same form as the output of
\code{\link[randomForestSRC]{rfsrc}}.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurementsTrain} and the same number of features
as it.}
}
\value{
For \code{rfsrcTrainInterface}, the trained Random Forest Survival model.
For \code{rfsrcPredictInterface}, a risk score prediction for each sample.
}
\description{
Random Forest Survival Model
}
