% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SelectParamsOrNULL-class}
\alias{SelectParamsOrNULL-class}
\alias{SelectParamsOrNULL}
\title{Union of A SelectParams Object and NULL}
\description{
Allows a slot to be either a SelectParams class object or empty. No
constructor.
}
\examples{

  ModellingParams(selectParams = NULL)
  ModellingParams(selectParams = SelectParams(differentMeansRanking))

}
\author{
Dario Strbenac
}
