\name{StageParamsOrMissing}
\docType{class}
\alias{StageParamsOrMissing}
\alias{StageParamsOrMissing-class}

\title{Union of A StageParams Object and NULL}
\description{
  Allows a slot to be either a class that has StageParams as its virtual parent class or empty. No constructor.
}

\author{Dario Strbenac}