\name{StageParamsOrMissingOrNULL}
\docType{class}
\alias{StageParamsOrMissingOrNULL}
\alias{StageParamsOrMissingOrNULL-class}

\title{Union of A StageParams Object and NULL}
\description{
  Allows a slot to be either a class that has StageParams as its virtual parent class or empty or NULL. No constructor.
}

\author{Dario Strbenac}