% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceSVM.R
\name{SVMtrainInterface}
\alias{SVMtrainInterface}
\alias{SVMtrainInterface,matrix-method}
\alias{SVMtrainInterface,DataFrame-method}
\alias{SVMtrainInterface,MultiAssayExperiment-method}
\alias{SVMpredictInterface}
\alias{SVMpredictInterface,svm,matrix-method}
\alias{SVMpredictInterface,svm,DataFrame-method}
\alias{SVMpredictInterface,svm,MultiAssayExperiment-method}
\title{An Interface for e1071 Package's Support Vector Machine Classifier.}
\usage{
SVMtrainInterface(measurementsTrain, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method (e.g. \code{verbose}) or options that are used by
the \code{\link[e1071]{svm}} function.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurementsTrain} and the same number of features
as it.}

\item{returnType}{Default: \code{"both"}. Either \code{"class"},
\code{"score"} or \code{"both"}. Sets the return value from the prediction
to either a vector of class labels, score for a sample belonging to the
second class, as determined by the factor levels, or both labels and scores
in a \code{data.frame}.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}

\item{model}{A fitted model as returned by \code{SVMtrainInterface}.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
For \code{SVMtrainInterface}, a trained SVM classifier of type
\code{svm}.  For \code{SVMpredictInterface}, either a factor vector of
predicted classes, a vector of secores for the second class, or a table of
both the class labels and second class scores, depending on the setting of
\code{returnType}.
}
\description{
\code{SVMtrainInterface} generates a trained SVM classifier and
\code{SVMpredictInterface} uses it to make predictions on a test data set.
}
\examples{

  if(require(e1071))
  {
    # Genes 76 to 100 have differential expression.
    genesMatrix <- sapply(1:100, function(sample) rnorm(25, 9, 0.3))
    genesMatrix <- rbind(genesMatrix, t(sapply(1:25, function(sample)
                                      c(rnorm(75, 9, 0.3), rnorm(25, 14, 0.3)))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    rownames(genesMatrix) <- paste("Sample", 1:nrow(genesMatrix))
    colnames(genesMatrix) <- paste("Gene", 1:ncol(genesMatrix))
    trainingSamples <- c(1:20, 26:45)
    testingSamples <- c(21:25, 46:50)
    
    classifier <- SVMtrainInterface(genesMatrix[trainingSamples, ],
                                     classes[trainingSamples], kernel = "linear")
    SVMpredictInterface(classifier, genesMatrix[testingSamples, ])
  }

}
\author{
Dario Strbenac
}
