% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_classifyr}
\alias{test_classifyr}
\title{Test the classifier model}
\usage{
test_classifyr(model, data)
}
\arguments{
\item{model}{A trained classifier.}

\item{data}{A data frame used for evaluation.}
}
\value{
A data frame with evaluation metrics.
}
\description{
Test the classifier model
}
