% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_classifyr}
\alias{train_classifyr}
\title{Train our example classifier model}
\usage{
train_classifyr(data, target, features = NULL)
}
\arguments{
\item{data}{A data frame with inputs and targets.}

\item{target}{A column name in data.}

\item{features}{Column names to use as features.
If \code{NULL}, use all non-target columns.}
}
\value{
A \code{classifyr_model} object.
}
\description{
Train our example classifier model
}
