% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TransformParams}
\alias{TransformParams}
\alias{TransformParams-class}
\alias{TransformParams,ANY-method}
\alias{TransformParams,function-method}
\alias{show,TransformParams-method}
\title{Parameters for Data Transformation}
\usage{
TransformParams(transform, ...)
}
\description{
Collects and checks necessary parameters required for transformation. The
empty constructor is for when no data transformation is desired. See
\code{\link{subtractFromLocation}} for an example of such a function.
}
\section{Constructor}{
 \describe{ \item{}{ \code{TransformParams(transform,
characteristics = DataFrame(), intermediate = character(0), ...)} Creates a
TransformParams object which stores the function which will do the
transformation and parameters that the function will use.  \describe{
\item{list("transform")}{A function which will do the transformation. The
first argument must be a \code{\link{DataFrame}} object.}
\item{list("characteristics")}{A \code{\link{DataFrame}} describing the
characteristics of data transformation to be done. First column must be
named \code{"charateristic"} and second column must be named \code{"value"}.
If using wrapper functions for data transformation in this package, the data
transformation name will automatically be generated and therefore it is not
necessary to specify it.} \item{list("intermediate")}{Character vector.
Names of any variables created in prior stages by \code{\link{runTest}} that
need to be passed to a feature selection function.} \item{list("...")}{Other
named parameters which will be used by the transformation function.} } } }
}

\examples{

  transformParams <- TransformParams(subtractFromLocation, location = "median")
  # Subtract all values from training set median, to obtain absolute deviations.

}
\author{
Dario Strbenac
}
