% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TransformParamsOrNULL-class}
\alias{TransformParamsOrNULL-class}
\alias{TransformParamsOrNULL}
\title{Union of A TransformParams Object and NULL}
\description{
Allows a slot to be either a TransformParams class object or empty. No
constructor.
}
\examples{

  ModellingParams(transformParams = NULL)
  ModellingParams(transformParams = TransformParams(subtractFromLocation),
                  selectParams = SelectParams(leveneRanking))

}
\author{
Dario Strbenac
}
