% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{List Available Feature Selection and Classification Approaches}
\usage{
available(what = c("classifier", "selectionMethod", "multiViewMethod"))
}
\arguments{
\item{what}{Default: \code{"classifier"}. Either \code{"classifier"}, \code{"selectionMethod"}
or \code{"multiViewMethod"}.}
}
\description{
Prints a list of keywords to use with \code{\link{crossValidate}}
}
\examples{

 available()
 
}
\author{
Dario Strbenac
}
