% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingCoxPH.R
\name{colCoxTests}
\alias{colCoxTests}
\title{A function to perform fast cox proportional hazard model tests}
\usage{
colCoxTests(X, y, option = c("fast", "slow"), ...)
}
\arguments{
\item{y}{matrix with first column as time and second column as event.}

\item{option}{whether to use the fast or slow method.}

\item{x}{matrix with variables as columns.}
}
\value{
CrossValParams object
}
\description{
A function to perform fast cox proportional hazard model tests
}
\examples{
data(asthma)
time <- rpois(nrow(measurements),100)
status <- sample(c(0,1), nrow(measurements), replace = TRUE)
x = measurements
y = cbind(time, status)
output <- colCoxTests(x, y, "fast")
}
