% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crissCrossValidate.R
\name{crissCrossPlot}
\alias{crissCrossPlot}
\title{A function to plot the output of the crissCrossValidate function.}
\usage{
crissCrossPlot(crissCrossResult, includeValues = FALSE)
}
\arguments{
\item{crissCrossResult}{The output of the crissCrossValidate function.}

\item{includeValues}{If TRUE, then the values of the matrix will be included in the plot.}
}
\description{
This function has been designed to give a heatmap output of the crissCrossValidate function.
}
\author{
Harry Robertson
}
