\name{DataFrameOrDataFrameList}
\docType{class}
\alias{DataFrameOrDataFrameList}
\alias{DataFrameOrDataFrameList-class}

\title{Union of A DataFrame and DataFrameList Class}
\description{
  Allows cross-validation to accept data as either a \code{DataFrame} (for a single data set) or \code{DataFrameList} (for a list of tables of related measurements, such as different projects measuring the same outcome and the same kind of measurements). No constructor.
}

\author{Dario Strbenac}