% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingDifferentMeans.R
\name{differentMeansRanking}
\alias{differentMeansRanking}
\alias{differentMeansRanking,matrix-method}
\alias{differentMeansRanking,DataFrame-method}
\alias{differentMeansRanking,MultiAssayExperiment-method}
\title{Ranking of Differentially Abundant Features}
\usage{
\S4method{differentMeansRanking}{matrix}(measurementsTrain, classesTrain, ...)

\S4method{differentMeansRanking}{DataFrame}(measurementsTrain, classesTrain, verbose = 3)

\S4method{differentMeansRanking}{MultiAssayExperiment}(measurementsTrain, targets = NULL, classesTrain, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}

\item{targets}{Names of data tables to be combined into a single table and
used in the analysis.}
}
\value{
A vector or data frame (if \code{MultiAssayExperiment} input) of
features, from the most promising features in the first position to the
least promising feature in the last position.
}
\description{
Uses an ordinary t-test if the data set has two classes or one-way ANOVA if
the data set has three or more classes to select differentially expressed
features.
}
\details{
This ranking method looks for changes in means and uses
\code{\link[genefilter]{rowttests}} to rank the features if there are two
classes or \code{\link[genefilter]{rowFtests}} if there are three or more
classes.
}
\examples{

    # Genes 76 to 100 have differential expression.
    genesMatrix <- sapply(1:100, function(sample) rnorm(25, 9, 0.3))
    genesMatrix <- rbind(genesMatrix, t(sapply(1:25, function(sample)
                                      c(rnorm(75, 9, 0.3), rnorm(25, 14, 0.3)))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    rownames(genesMatrix) <- paste("Sample", 1:50)
    colnames(genesMatrix) <- paste("Gene", 1:100)
    
    ranked <- differentMeansRanking(genesMatrix, classes)
    head(ranked)
}
\author{
Dario Strbenac
}
