% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingDMD.R
\name{DMDranking}
\alias{DMDranking}
\alias{DMDranking,matrix-method}
\alias{DMDranking,DataFrame-method}
\alias{DMDranking,MultiAssayExperiment-method}
\title{Ranking by Differential Distributions with Differences in Means or Medians
and a Deviation Measure}
\usage{
DMDranking(measurements, ...)
}
\arguments{
\item{measurements}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix}, the rows are features, and the columns are samples.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method or parameters for
\code{\link{getLocationsAndScales}}, such as \code{location}, \code{scale}.}

\item{classes}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurements} if it is a
\code{\link{matrix}} (i.e. number of columns) or a \code{\link{DataFrame}}
(i.e. number of rows) or a character vector of length 1 containing the
column name in \code{measurements} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurements)} if \code{measurements} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"clinical"} is also a valid value
and specifies that numeric variables from the clinical data table will be
used.}

\item{differences}{Default: \code{"both"}. Either \code{"both"},
\code{"location"}, or \code{"scale"}.  The type of differences to consider.
If both are considered then the absolute difference in location and the
absolute difference in scale are summed.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give.  This function only prints progress messages if
the value is 3.}
}
\value{
A vector of feature indices, from the most promising features in the
first position to the least promising feature in the last position.
}
\description{
Ranks features by largest Differences in Means/Medians and Deviations.
}
\details{
DMD is defined as \eqn{\sum_{i = 1}\sum_{j = i + 1}|location_i - location_j|
+ |scale_i - scale_j|}{sum(|location i - location j|) + sum(|scale i - scale
j|), i not j, i < j}. The subscripts denote the class for which the
parameter is calculated for.
}
\examples{

  # First 20 features have bimodal distribution for Poor class.
  # Other 80 features have normal distribution for both classes.
  genesMatrix <- sapply(1:25, function(sample)
                              {
                                randomMeans <- sample(c(8, 12), 20, replace = TRUE)
                                c(rnorm(20, randomMeans, 1), rnorm(80, 10, 1))
                              }
                       )
  genesMatrix <- cbind(genesMatrix, sapply(1:25, function(sample) rnorm(100, 10, 1)))
  classes <- factor(rep(c("Poor", "Good"), each = 25))
  
  ranked <- DMDranking(genesMatrix, classes)
  head(ranked)

}
\author{
Dario Strbenac
}
