% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceElasticNetGLM.R
\name{elasticNetFeatures}
\alias{elasticNetFeatures}
\alias{elasticNetFeatures,multnet-method}
\title{Extract Vectors of Ranked and Selected Features From an Elastic Net GLM
Object}
\usage{
elasticNetFeatures(model, ...)
}
\arguments{
\item{model}{A fitted multinomial GLM which was created by
\code{\link[glmnet]{glmnet}}.}
}
\value{
An \code{list} object. The first element is a vector or data frame
of ranked features, the second is a vector or data frame of selected
features.
}
\description{
Provides a ranking of features based on the magnitude of fitted GLM
coefficients. Also provides the selected features which are those with a
non-zero coefficient.
}
\examples{

    if(require(glmnet))
    {
      # Genes 76 to 100 have differential expression.
    genesMatrix <- sapply(1:100, function(sample) rnorm(25, 9, 0.3))
    genesMatrix <- rbind(genesMatrix, t(sapply(1:25, function(sample)
                                      c(rnorm(75, 9, 0.3), rnorm(25, 14, 0.3)))))
    classes <- factor(rep(c("Poor", "Good"), each = 25))
    rownames(genesMatrix) <- paste("Sample", 1:nrow(genesMatrix))
    colnames(genesMatrix) <- paste("Gene", 1:ncol(genesMatrix))
                                             
      # alpha is a user-specified tuning parameter.
      # lambda is automatically tuned, based on glmnet defaults, if not user-specified.                 
      CVparams <- CrossValParams("k-Fold")
      
      trainParams <- TrainParams(elasticNetGLMtrainInterface, nlambda = 500)
      predictParams <- PredictParams(elasticNetGLMpredictInterface)
      modParams <- ModellingParams(selectParams = NULL, trainParams = trainParams,
                                   predictParams = predictParams)
                                   
      classified <- runTests(genesMatrix, classes, CVparams, modParams)
                                        
      elasticNetFeatures(models(classified)[[1]])
    }

}
\author{
Dario Strbenac
}
