% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFisherDiscriminant.R
\name{fisherDiscriminant}
\alias{fisherDiscriminant}
\alias{fisherDiscriminant,matrix-method}
\alias{fisherDiscriminant,DataFrame-method}
\alias{fisherDiscriminant,MultiAssayExperiment-method}
\title{Classification Using Fisher's LDA}
\usage{
\S4method{fisherDiscriminant}{matrix}(measurementsTrain, classesTrain, measurementsTest, ...)

\S4method{fisherDiscriminant}{DataFrame}(
  measurementsTrain,
  classesTrain,
  measurementsTest,
  returnType = c("both", "class", "score"),
  verbose = 3
)

\S4method{fisherDiscriminant}{MultiAssayExperiment}(
  measurementsTrain,
  measurementsTest,
  targets = names(measurementsTrain),
  classesTrain,
  ...
)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method.}

\item{classesTrain}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurementsTrain} if it is a
\code{\link{matrix}} or a \code{\link{DataFrame}} or a character vector of length 1
containing the column name in \code{measurementsTrain} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurementsTrain)} if \code{measurementsTrain} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{measurementsTest}{An object of the same class as \code{measurementsTrain} with no
samples in common with \code{measurementsTrain} and the same number of features
as it.}

\item{returnType}{Default: \code{"both"}. Either \code{"class"},
\code{"score"}, or \code{"both"}.  Sets the return value from the prediction
to either a vector of class labels, score for a sample belonging to the
second class, as determined by the factor levels, or both labels and scores
in a \code{data.frame}.}

\item{verbose}{Default: 3. A number between 0 and 3 for the amount of
progress messages to give. This function only prints progress messages if
the value is 3.}

\item{targets}{If \code{measurementsTrain} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}
}
\value{
A vector or \code{data.frame} of class prediction information, as
long as the number of samples in the test data.
}
\description{
Finds the decision boundary using the training set, and gives predictions
for the test set. Unlike ordinary LDA, Fisher's version does not have assumptions
about the normality of the features. Data tables which consist entirely of non-numeric
data cannot be analysed.
}
\examples{

  trainMatrix <- matrix(rnorm(1000, 8, 2), nrow = 10)
  classes <- factor(rep(c("Poor", "Good"), each = 5))
  
  # Make first 30 genes increased in value for poor samples.
  trainMatrix[1:5, 1:30] <- trainMatrix[1:5, 1:30] + 5
  
  testMatrix <- matrix(rnorm(1000, 8, 2), nrow = 10)
  
  # Make first 30 genes increased in value for sixth to tenth samples.
  testMatrix[6:10, 1:30] <- testMatrix[6:10, 1:30] + 5
  
  fisherDiscriminant(trainMatrix, classes, testMatrix)

}
\author{
Dario Strbenac
}
