% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METABRICclinical}
\alias{METABRICclinical}
\alias{clinical}
\title{METABRIC Clinical Data}
\format{
\code{clinical} A \code{\link{DataFrame}} containing clinical data.
}
\source{
Dynamics of Breast Cancer Relapse Reveal Late-recurring ER-positive Genomic Subgroups, \emph{Nature},
2019.  Webpage: \url{https://www.nature.com/articles/s43018-020-0026-6}
}
\description{
470 patients with eight features.
}
\keyword{datasets}
