% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,PredictParams-method}
\alias{show,PredictParams-method}
\title{Inspect Prediction Function Details}
\usage{
\S4method{show}{PredictParams}(object)
}
\arguments{
\item{object}{An object of class \code{TrainParams} to inspect.}
}
\description{
Inspect Prediction Function Details
}
