% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PredictParamsOrNULL}
\alias{PredictParamsOrNULL}
\alias{PredictParamsOrNULL-class}
\title{Union of A PredictParams Object and NULL}
\description{
Allows a slot to be either a PredictParams class object or empty. No
constructor. In other words, the training function specified also makes
predictions with the test set.
}
\examples{

  ModellingParams(trainParams = TrainParams(kNNinterface, k = 5), predictParams = NULL)

}
\author{
Dario Strbenac
}
