% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\alias{prepareData,matrix-method}
\alias{prepareData,DataFrame-method}
\alias{prepareData,MultiAssayExperiment-method}
\title{Convert Different Data Classes into DataFrame and Filter Features}
\usage{
\S4method{prepareData}{matrix}(measurements, outcome, ...)
}
\arguments{
\item{measurements}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing all of the data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns
are features.}

\item{...}{Variables not used by the \code{matrix} nor the
\code{MultiAssayExperiment} method which are passed into and used by the
\code{DataFrame} method.}

\item{outcome}{Either a factor vector of classes, a \code{\link{Surv}} object, or
a character string, or vector of such strings, containing column name(s) of column(s)
containing either classes or time and event information about survival.}

\item{outcomeColumns}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the column (class) or columns (survival) in the table extracted by \code{colData(data)}
that contain(s) the each individual's outcome to use for prediction.}

\item{useFeatures}{If \code{measurements} is a \code{MultiAssayExperiment},
a two-column table of features to use. The first column must have assay names
and the second column must have feature names found for that assay. \code{"clinical"} is
also a valid assay name and refers to the clinical data table. \code{"all"} is a special
keyword that means all features (passing any other filters) of that assay will be used 
for modelling. Otherwise, a character vector of feature names to use suffices.}

\item{maxMissingProp}{Default: 0.0. A proportion less than 1 which is the maximum
tolerated proportion of missingness for a feature to be retained for modelling.}

\item{topNvariance}{Default: NULL. An integer number of most variable features to subset to.}
}
\value{
A list of length two. The first element is a \code{\link{DataFrame}} of features
and the second element is the outcomes to use for modelling.
}
\description{
Input data could be of matrix, MultiAssayExperiment, or DataFrame format and this
function will prepare a DataFrame of features and a vector of outcomes and help
to exclude nuisance features such as dates or unique sample identifiers from
future modelling.
}
\author{
Dario Strbenac
}
