% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickClassify.R
\name{quickClassify}
\alias{quickClassify}
\alias{quickClassify,matrix-method}
\alias{quickClassify,DataFrameOrDataFrameList-method}
\alias{quickClassify,MultiAssayExperiment-method}
\title{Quick and Easy Cross-validated Classification}
\usage{
quickClassify(measurements, ...)
}
\arguments{
\item{measurements}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data.  For a
\code{matrix}, the rows are features, and the columns are samples.}

\item{classes}{A vector of class labels of class \code{\link{factor}} of the
same length as the number of samples in \code{measurements} if it is a
\code{\link{matrix}} (i.e. number of columns) or a \code{\link{DataFrame}}
(i.e. number of rows) or a character vector of length 1 containing the
column name in \code{measurements} if it is a \code{\link{DataFrame}} or the
column name in \code{colData(measurements)} if \code{measurements} is a
\code{\link{MultiAssayExperiment}}. If a column name, that column will be
removed before training.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"clinical"} is also a valid value
and specifies that the clinical data table will be used.}

\item{permutations}{Default: 100. Number of times to permute the samples
before folding them. Set to \code{NULL} for a singlee k-fold
cross-validation.}

\item{folds}{Default: 5. The number of disjoint sets to partition the
samples into. Set this value to the number of features in
\code{measurements} (if it is not a list of data sets of unequal sample
numbers) to effectively perform a traditional leave-one-out
cross-validation. This variable must be some integer.}

\item{cores}{Default: 1. Number of CPU cores to use. Each cross-validation
iteration is done on one core.}

\item{classifier}{A short name describing the kind of classification to do.
A vector means that each classification will be applied to the data in turn.
Specify one name if only one classifier desired. The options represent:
\itemize{\item \code{"DLDA"}: Feature selection based on differences in
means ranking and diagonal LDA classifier.  \item \code{"FishersLDA"}:
Transformation by subtraction of values from the mean value of all samples
in the training set followed by selection by Levene's test statistic ranking
based on differential variability and classification by Fisher's LDA.  \item
\code{"NSC"}: Nearest shrunken centroid classifier. No explicit feature
selection before classifier.  \item \code{"naiveBayes"}: Feature selection
based on differences in median and Qn followed by naive Bayes classifier
using weighted voting based on the difference between class densities. See
\code{\link{DMDranking}} for more details. Good for differential
dsitribution classification.  \item \code{"elasticNet"}: Elastic Net GLM
classifier with \code{"multinomial"} as the family.  \item \code{"SVM"}:
Polynomial kernel support vector machine. Does parameter tuning of poly
nomial degree between 2 and 8 and the cost penality between 0.00001 to
100000 in power increments of 1.  \item \code{"randomForest"}: Random forest
with the number of variables tried at each split eing 40\% of the total
number of variables in the data set.  \item \code{"kTSP"}: Feature selection
based on pairs differences ranking and classification based on weighted
voting of the selected feature pairs.  } \item \code{datasetMode} Default: "each".
Whether to classify each input data set individually or concatenate
different ones into a single table by specifying \code{"combine"} or do
classification on all possible pairs of data sets, using the first data set
of the pair for training and the second of the pair for making predictions
with by specifying \code{"allPairs"}. Only relevant if \code{measurements}
is a \code{DataFrame} created by the \code{MultiAssayExperiment} method,
possibly containing variables from multiple assays or a
\code{DataFrameList}.  \item...Either variables not used by the
\code{matrix} nor the \code{MultiAssayExperiment} method which are passed
into and used by the \code{DataFrame} method or variables not used by
\code{quickClassify} but used by runTests (i.e. \code{characteristics} and
\code{verbose}.}
}
\description{
This convenience function allows specification of cross-validation and
classification by a couple of pre-defined keywords and all parameter setting
is automatically taken care of. This avoids the need to create any S4
parameter objects.
}
