% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,SelectParams-method}
\alias{show,SelectParams-method}
\title{Container for Storing Details of Feature Selection Function(s)}
\usage{
\S4method{show}{SelectParams}(object)
}
\arguments{
\item{object}{An object of class \code{SelectParams} to inspect.}
}
\description{
Container for Storing Details of Feature Selection Function(s)
}
