% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{StageParamsOrMissing}
\alias{StageParamsOrMissing}
\alias{StageParamsOrMissing-class}
\alias{StageParamsOrMissingOrNULL}
\alias{StageParamsOrMissingOrNULL-class}
\title{Union of A StageParams Object and NULL}
\description{
StageParamsOrMissing: Allows a slot to be either a class that has StageParams as its virtual
parent class or empty. No constructor.
StageParamsOrMissingOrNULL: Allows a slot to be either a class that has StageParams as its virtual
parent class or empty or NULL. No constructor.
}
\author{
Dario Strbenac
}
