% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractFromLocation.R
\name{subtractFromLocation}
\alias{subtractFromLocation}
\alias{subtractFromLocation,matrix,matrix-method}
\alias{subtractFromLocation,DataFrame,DataFrame-method}
\alias{subtractFromLocation,MultiAssayExperiment,MultiAssayExperiment-method}
\title{Subtract Numeric Feature Measurements from a Location}
\usage{
subtractFromLocation(measurementsTrain, measurementsTest, ...)
}
\arguments{
\item{measurementsTrain}{Either a \code{\link{matrix}}, \code{\link{DataFrame}}
or \code{\link{MultiAssayExperiment}} containing the training data. For a
\code{matrix} or \code{\link{DataFrame}}, the rows are samples, and the columns are features.
If of type \code{\link{DataFrame}} or \code{\link{MultiAssayExperiment}}, the data set is subset
to only those features of type \code{numeric}.}

\item{measurementsTest}{A data set of the same type as \code{measurementsTrain} with no samples in common with it.
The subtraction will also be performed to it.}

\item{location}{Character. Either "mean" or "median".}

\item{absolute}{Logical. Default: \code{TRUE}. If \code{TRUE}, then absolute
values of the differences are returned. Otherwise, they are signed.}

\item{targets}{If \code{measurements} is a \code{MultiAssayExperiment}, the
names of the data tables to be used. \code{"sampleInfo"} is also a valid value
and specifies that numeric variables from the sample information data table will be
used.}

\item{verbose}{Default: 3. A progress message is shown if this value is 3.}
}
\value{
The same class of variable as the input variable \code{measurements}
is, with the numeric features subtracted from the calculated location.
}
\description{
For each numeric feature, calculates the location, and subtracts all
measurements from that location.
}
\details{
Only the samples specified by \code{measurementsTrain} are used in the calculation of
the location.
}
\examples{

  aMatrix <- matrix(1:100, ncol = 10)
  subtractFromLocation(aMatrix[1:5,], aMatrix[6:10, ], "median")

}
\author{
Dario Strbenac
}
