% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,TrainParams-method}
\alias{show,TrainParams-method}
\title{Inspect Model Training Details}
\usage{
\S4method{show}{TrainParams}(object)
}
\arguments{
\item{object}{An object of class \code{TrainParams} to inspect.}
}
\description{
Inspect Model Training Details
}
